bettered some examples
Former-commit-id: eb15593f1e9560bfa3377c269aa0e5c6112b5484

+++ Examples/src/RK45/ButcherRKF.hpp
@@ -32,7 +32,7 @@ struct ButcherArray{
   ButcherArray(Atable const &a,
                std::array<double,NSTAGES> const & b1,
                std::array<double,NSTAGES> const & b2,
-               double ord):
+               int ord):
     A(a),b1(b1),b2(b2),order(ord)
   {
     c.fill(0.0);
